+++ Braintree/src/androidTest/java/com/braintreepayments/api/GooglePaymentTest.java
@@ -5,16 +5,18 @@ import android.content.Intent;
 import android.os.Bundle;
 import android.support.test.runner.AndroidJUnit4;
 
+import com.braintreepayments.api.exceptions.InvalidArgumentException;
 import com.braintreepayments.api.interfaces.BraintreeResponseListener;
 import com.braintreepayments.api.interfaces.ConfigurationListener;
 import com.braintreepayments.api.interfaces.TokenizationParametersListener;
+import com.braintreepayments.api.models.Authorization;
 import com.braintreepayments.api.models.BraintreeRequestCodes;
 import com.braintreepayments.api.models.Configuration;
 import com.braintreepayments.api.models.GooglePaymentRequest;
 import com.braintreepayments.api.test.BraintreeActivityTestRule;
 import com.braintreepayments.api.test.TestActivity;
 import com.braintreepayments.testutils.TestConfigurationBuilder;
-import com.braintreepayments.testutils.TestConfigurationBuilder.TestAndroidPayConfigurationBuilder;
+import com.braintreepayments.testutils.TestConfigurationBuilder.TestGooglePaymentConfigurationBuilder;
 import com.google.android.gms.wallet.CardRequirements;
 import com.google.android.gms.wallet.PaymentDataRequest;
 import com.google.android.gms.wallet.PaymentMethodTokenizationParameters;
@@ -23,9 +25,11 @@ import com.google.android.gms.wallet.TransactionInfo;
 import com.google.android.gms.wallet.WalletConstants;
 import com.google.android.gms.wallet.WalletConstants.CardNetwork;
 
+import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
 import org.junit.Before;
+import org.junit.Ignore;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -33,7 +37,6 @@ import org.mockito.ArgumentCaptor;
 import org.mockito.InOrder;
 
 import java.util.Collection;
-import java.util.List;
 import java.util.concurrent.CountDownLatch;
 
 import static com.braintreepayments.api.BraintreeFragmentTestUtils.getFragment;
@@ -48,9 +51,11 @@ import static junit.framework.Assert.assertNotNull;
 import static junit.framework.Assert.assertNull;
 import static junit.framework.Assert.assertTrue;
 import static junit.framework.Assert.fail;
+import static org.mockito.Matchers.any;
 import static org.mockito.Matchers.eq;
 import static org.mockito.Mockito.inOrder;
 import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.when;
 
 @RunWith(AndroidJUnit4.class)
 public class GooglePaymentTest {
@@ -66,7 +71,7 @@ public class GooglePaymentTest {
     public void setup() {
         mLatch = new CountDownLatch(1);
         mBaseConfiguration = new TestConfigurationBuilder()
-                .androidPay(new TestAndroidPayConfigurationBuilder()
+                .googlePayment(new TestGooglePaymentConfigurationBuilder()
                         .googleAuthorizationFingerprint("google-auth-fingerprint"))
                 .merchantId("android-pay-merchant-id");
     }
@@ -74,7 +79,7 @@ public class GooglePaymentTest {
     @Test(timeout = 5000)
     public void isReadyToPay_returnsFalseWhenAndroidPayIsNotEnabled() throws Exception {
         String configuration = new TestConfigurationBuilder()
-                .androidPay(new TestAndroidPayConfigurationBuilder().enabled(false))
+                .googlePayment(new TestGooglePaymentConfigurationBuilder().enabled(false))
                 .build();
 
         BraintreeFragment fragment = getFragment(mActivityTestRule.getActivity(), TOKENIZATION_KEY, configuration);
@@ -102,41 +107,11 @@ public class GooglePaymentTest {
 
         GooglePayment.requestPayment(fragment, googlePaymentRequest);
 
-        ArgumentCaptor<Intent> captor = ArgumentCaptor.forClass(Intent.class);
-        verify(fragment).startActivityForResult(captor.capture(), eq(BraintreeRequestCodes.GOOGLE_PAYMENT));
-        Intent intent = captor.getValue();
-        assertEquals(GooglePaymentActivity.class.getName(), intent.getComponent().getClassName());
-        assertEquals(WalletConstants.ENVIRONMENT_TEST, intent.getIntExtra(EXTRA_ENVIRONMENT, -1));
-        PaymentDataRequest paymentDataRequest = intent.getParcelableExtra(EXTRA_PAYMENT_DATA_REQUEST);
-        assertEquals(googlePaymentRequest.getTransactionInfo(), paymentDataRequest.getTransactionInfo());
-        assertEquals(2, paymentDataRequest.getAllowedPaymentMethods().size());
-        assertTrue(paymentDataRequest.getAllowedPaymentMethods().contains(WalletConstants.PAYMENT_METHOD_CARD));
-        assertTrue(paymentDataRequest.getAllowedPaymentMethods().contains(WalletConstants.PAYMENT_METHOD_TOKENIZED_CARD));
-        assertFalse(paymentDataRequest.isEmailRequired());
-        assertFalse(paymentDataRequest.isPhoneNumberRequired());
-        List<Integer> allowedCardNetworks = paymentDataRequest.getCardRequirements().getAllowedCardNetworks();
-        assertEquals(4, allowedCardNetworks.size());
-        assertTrue(allowedCardNetworks.contains(WalletConstants.CARD_NETWORK_VISA));
-        assertTrue(allowedCardNetworks.contains(WalletConstants.CARD_NETWORK_DISCOVER));
-        assertTrue(allowedCardNetworks.contains(WalletConstants.CARD_NETWORK_MASTERCARD));
-        assertTrue(allowedCardNetworks.contains(WalletConstants.CARD_NETWORK_AMEX));
-        assertEquals(WalletConstants.PAYMENT_METHOD_TOKENIZATION_TYPE_PAYMENT_GATEWAY,
-                paymentDataRequest.getPaymentMethodTokenizationParameters().getPaymentMethodTokenizationType());
-
-        Bundle expectedParameters = GooglePayment.getTokenizationParameters(fragment)
-                .getParameters();
-        Bundle actualParameters = paymentDataRequest.getPaymentMethodTokenizationParameters().getParameters();
-        assertEquals(expectedParameters.getString("gateway"), actualParameters.getString("gateway"));
-        assertEquals(expectedParameters.getString("braintree:merchantId"), actualParameters.getString("braintree:merchantId"));
-        assertEquals(expectedParameters.getString("braintree:authorizationFingerprint"),
-                actualParameters.getString("braintree:authorizationFingerprint"));
-        assertEquals(expectedParameters.getString("braintree:apiVersion"), actualParameters.getString("braintree:apiVersion"));
-        assertEquals(expectedParameters.getString("braintree:sdkVersion"), actualParameters.getString("braintree:sdkVersion"));
-        assertEquals(expectedParameters.getString("braintree:metadata"), actualParameters.getString("braintree:metadata"));
+        verify(fragment).startActivityForResult(any(Intent.class), eq(BraintreeRequestCodes.GOOGLE_PAYMENT));
     }
 
     @Test
-    public void requestPayment_startsActivityWithOptionalValues() {
+    public void requestPayment_startsActivityWithOptionalValues_GooglePaymentV1() {
         BraintreeFragment fragment = getSetupFragment();
         GooglePaymentRequest googlePaymentRequest = new GooglePaymentRequest()
                 .allowPrepaidCards(true)
@@ -171,6 +146,100 @@ public class GooglePaymentTest {
         assertNotNull(paymentDataRequest.getShippingAddressRequirements().getAllowedCountryCodes());
         assertTrue(paymentDataRequest.getShippingAddressRequirements().getAllowedCountryCodes().contains("USA"));
         assertTrue(paymentDataRequest.isUiRequired());
+    }
+
+    @Ignore("Requires Google-Payment@2+")
+    @Test
+    public void requestPayment_startsActivityWithOptionalValues_GooglePaymentV2() throws JSONException {
+        BraintreeFragment fragment = getSetupFragment();
+        GooglePaymentRequest googlePaymentRequest = new GooglePaymentRequest()
+                .allowPrepaidCards(true)
+                .billingAddressFormat(1)
+                .billingAddressRequired(true)
+                .emailRequired(true)
+                .phoneNumberRequired(true)
+                .shippingAddressRequired(true)
+                .shippingAddressRequirements(ShippingAddressRequirements.newBuilder().addAllowedCountryCode("USA").build())
+                .transactionInfo(TransactionInfo.newBuilder()
+                        .setTotalPrice("1.00")
+                        .setTotalPriceStatus(WalletConstants.TOTAL_PRICE_STATUS_FINAL)
+                        .setCurrencyCode("USD")
+                        .build());
+
+        GooglePayment.requestPayment(fragment, googlePaymentRequest);
+
+        ArgumentCaptor<Intent> captor = ArgumentCaptor.forClass(Intent.class);
+        verify(fragment).startActivityForResult(captor.capture(), eq(BraintreeRequestCodes.GOOGLE_PAYMENT));
+        Intent intent = captor.getValue();
+
+        assertEquals(GooglePaymentActivity.class.getName(), intent.getComponent().getClassName());
+        assertEquals(WalletConstants.ENVIRONMENT_TEST, intent.getIntExtra(EXTRA_ENVIRONMENT, -1));
+        PaymentDataRequest paymentDataRequest = intent.getParcelableExtra(EXTRA_PAYMENT_DATA_REQUEST);
+
+        JSONObject paymentDataRequestJson = new JSONObject(paymentDataRequest.toJson());
+
+        assertEquals(2, paymentDataRequestJson.get("apiVersion"));
+        assertEquals(0, paymentDataRequestJson.get("apiVersionMinor"));
+
+        assertEquals(true, paymentDataRequestJson.get("emailRequired"));
+        assertEquals(true, paymentDataRequestJson.get("shippingAddressRequired"));
+
+        JSONObject transactionInfoJson = paymentDataRequestJson.getJSONObject("transactionInfo");
+        assertEquals("FINAL", transactionInfoJson.getString("totalPriceStatus"));
+        assertEquals("1.00", transactionInfoJson.getString("totalPrice"));
+        assertEquals("USD", transactionInfoJson.getString("currencyCode"));
+
+        JSONArray allowedPaymentMethods = paymentDataRequestJson.getJSONArray("allowedPaymentMethods");
+        JSONObject paypal = allowedPaymentMethods.getJSONObject(0);
+        assertEquals("PAYPAL", paypal.getString("type"));
+
+        JSONArray purchaseUnits = paypal.getJSONObject("parameters")
+                .getJSONObject("purchase_context")
+                .getJSONArray("purchase_units");
+        assertEquals(1, purchaseUnits.length());
+
+        JSONObject purchaseUnit = purchaseUnits.getJSONObject(0);
+        assertEquals("paypal-client-id", purchaseUnit.getJSONObject("payee")
+                .getString("client_id"));
+        assertEquals("true", purchaseUnit.getString("recurring_payment"));
+
+        JSONObject paypalTokenizationSpecification = paypal.getJSONObject("tokenizationSpecification");
+        assertEquals("PAYMENT_GATEWAY", paypalTokenizationSpecification.getString("type"));
+
+        JSONObject paypalTokenizationSpecificationParams = paypalTokenizationSpecification.getJSONObject("parameters");
+        assertEquals("braintree", paypalTokenizationSpecificationParams.getString("gateway"));
+        assertEquals("v1", paypalTokenizationSpecificationParams.getString("braintree:apiVersion"));
+        assertEquals(BuildConfig.VERSION_NAME, paypalTokenizationSpecificationParams.getString("braintree:sdkVersion"));
+        assertEquals("android-pay-merchant-id", paypalTokenizationSpecificationParams.getString("braintree:merchantId"));
+        assertEquals("{\"source\":\"client\",\"version\":\"" + BuildConfig.VERSION_NAME+ "\",\"platform\":\"android\"}", paypalTokenizationSpecificationParams.getString("braintree:metadata"));
+        assertFalse(paypalTokenizationSpecificationParams.has("braintree:clientKey"));
+        assertEquals("paypal-client-id", paypalTokenizationSpecificationParams.getString("braintree:paypalClientId"));
+
+        JSONObject card = allowedPaymentMethods.getJSONObject(1);
+        assertEquals("CARD", card.getString("type"));
+
+        JSONObject cardParams = card.getJSONObject("parameters");
+        assertTrue(cardParams.getBoolean("billingAddressRequired"));
+        assertTrue(cardParams.getBoolean("allowPrepaidCards"));
+
+        assertEquals("PAN_ONLY", cardParams.getJSONArray("allowedAuthMethods").getString(0));
+        assertEquals("CRYPTOGRAM_3DS", cardParams.getJSONArray("allowedAuthMethods").getString(1));
+
+        assertEquals("VISA", cardParams.getJSONArray("allowedCardNetworks").getString(0));
+        assertEquals("MASTERCARD", cardParams.getJSONArray("allowedCardNetworks").getString(1));
+        assertEquals("AMEX", cardParams.getJSONArray("allowedCardNetworks").getString(2));
+        assertEquals("DISCOVER", cardParams.getJSONArray("allowedCardNetworks").getString(3));
+
+        JSONObject tokenizationSpecification = card.getJSONObject("tokenizationSpecification");
+        assertEquals("PAYMENT_GATEWAY", tokenizationSpecification.getString("type"));
+
+        JSONObject cardTokenizationSpecificationParams = tokenizationSpecification.getJSONObject("parameters");
+        assertEquals("braintree", cardTokenizationSpecificationParams.getString("gateway"));
+        assertEquals("v1", cardTokenizationSpecificationParams.getString("braintree:apiVersion"));
+        assertEquals(BuildConfig.VERSION_NAME, cardTokenizationSpecificationParams.getString("braintree:sdkVersion"));
+        assertEquals("android-pay-merchant-id", cardTokenizationSpecificationParams.getString("braintree:merchantId"));
+        assertEquals("{\"source\":\"client\",\"version\":\"" + BuildConfig.VERSION_NAME+ "\",\"platform\":\"android\"}", cardTokenizationSpecificationParams.getString("braintree:metadata"));
+        assertEquals("sandbox_tokenization_key", cardTokenizationSpecificationParams.getString("braintree:clientKey"));
     }
 
     @Test
@@ -241,7 +310,7 @@ public class GooglePaymentTest {
 
                 assertEquals("braintree", tokenizationParameters.getString("gateway"));
                 assertEquals(configuration.getMerchantId(), tokenizationParameters.getString("braintree:merchantId"));
-                assertEquals(configuration.getAndroidPay().getGoogleAuthorizationFingerprint(),
+                assertEquals(configuration.getGooglePayment().getGoogleAuthorizationFingerprint(),
                         tokenizationParameters.getString("braintree:authorizationFingerprint"));
                 assertEquals("v1", tokenizationParameters.getString("braintree:apiVersion"));
                 assertEquals(BuildConfig.VERSION_NAME, tokenizationParameters.getString("braintree:sdkVersion"));
@@ -293,7 +362,7 @@ public class GooglePaymentTest {
 
     @Test(timeout = 5000)
     public void getAllowedCardNetworks_returnsSupportedNetworks() throws InterruptedException {
-        String configuration = mBaseConfiguration.androidPay(mBaseConfiguration.androidPay()
+        String configuration = mBaseConfiguration.googlePayment(mBaseConfiguration.googlePayment()
                 .supportedNetworks(new String[]{"visa", "mastercard", "amex", "discover"}))
                 .build();
 
@@ -319,7 +388,7 @@ public class GooglePaymentTest {
 
     @Test(timeout = 5000)
     public void getTokenizationParameters_returnsCorrectParametersInCallback() throws Exception {
-        String config = mBaseConfiguration.androidPay(mBaseConfiguration.androidPay()
+        String config = mBaseConfiguration.googlePayment(mBaseConfiguration.googlePayment()
                 .supportedNetworks(new String[]{"visa", "mastercard", "amex", "discover"}))
                 .build();
         final Configuration configuration = Configuration.fromJson(config);
@@ -333,7 +402,7 @@ public class GooglePaymentTest {
                 assertEquals("braintree", parameters.getParameters().getString("gateway"));
                 assertEquals(configuration.getMerchantId(),
                         parameters.getParameters().getString("braintree:merchantId"));
-                assertEquals(configuration.getAndroidPay().getGoogleAuthorizationFingerprint(),
+                assertEquals(configuration.getGooglePayment().getGoogleAuthorizationFingerprint(),
                         parameters.getParameters().getString("braintree:authorizationFingerprint"));
                 assertEquals("v1",
                         parameters.getParameters().getString("braintree:apiVersion"));
@@ -365,13 +434,24 @@ public class GooglePaymentTest {
     }
 
     private BraintreeFragment getSetupFragment() {
-        String configuration = mBaseConfiguration.androidPay(mBaseConfiguration.androidPay()
+        String configuration = mBaseConfiguration.googlePayment(mBaseConfiguration.googlePayment()
+                .environment("sandbox")
                 .supportedNetworks(new String[]{"visa", "mastercard", "amex", "discover"}))
+                .paypal(new TestConfigurationBuilder.TestPayPalConfigurationBuilder(true)
+                        .clientId("paypal-client-id"))
                 .withAnalytics()
                 .build();
 
-        return new MockFragmentBuilder()
+        BraintreeFragment fragment = new MockFragmentBuilder()
                 .configuration(configuration)
                 .build();
+
+        try {
+            when(fragment.getAuthorization()).thenReturn(Authorization.fromString("sandbox_tokenization_key"));
+        } catch (InvalidArgumentException e) {
+            throw new RuntimeException(e);
+        }
+
+        return fragment;
     }
 }