+++ Braintree/src/main/java/com/braintreepayments/api/models/Configuration.java
@@ -28,7 +28,7 @@ public class Configuration {
     private static final String PAYPAL_ENABLED_KEY = "paypalEnabled";
     private static final String PAYPAL_KEY = "paypal";
     private static final String KOUNT_KEY = "kount";
-    private static final String ANDROID_PAY_KEY = "androidPay";
+    private static final String GOOGLE_PAYMENT_KEY = "androidPay";
     private static final String THREE_D_SECURE_ENABLED_KEY = "threeDSecureEnabled";
     private static final String PAY_WITH_VENMO_KEY = "payWithVenmo";
     private static final String UNIONPAY_KEY = "unionPay";
@@ -51,7 +51,7 @@ public class Configuration {
     private CardConfiguration mCardConfiguration;
     private boolean mPaypalEnabled;
     private PayPalConfiguration mPayPalConfiguration;
-    private AndroidPayConfiguration mAndroidPayConfiguration;
+    private GooglePaymentConfiguration mGooglePaymentConfiguration;
     private boolean mThreeDSecureEnabled;
     private VenmoConfiguration mVenmoConfiguration;
     private KountConfiguration mKountConfiguration;
@@ -89,7 +89,7 @@ public class Configuration {
         mCardConfiguration = CardConfiguration.fromJson(json.optJSONObject(CARD_KEY));
         mPaypalEnabled = json.optBoolean(PAYPAL_ENABLED_KEY, false);
         mPayPalConfiguration = PayPalConfiguration.fromJson(json.optJSONObject(PAYPAL_KEY));
-        mAndroidPayConfiguration = AndroidPayConfiguration.fromJson(json.optJSONObject(ANDROID_PAY_KEY));
+        mGooglePaymentConfiguration = GooglePaymentConfiguration.fromJson(json.optJSONObject(GOOGLE_PAYMENT_KEY));
         mThreeDSecureEnabled = json.optBoolean(THREE_D_SECURE_ENABLED_KEY, false);
         mVenmoConfiguration = VenmoConfiguration.fromJson(json.optJSONObject(PAY_WITH_VENMO_KEY));
         mKountConfiguration = KountConfiguration.fromJson(json.optJSONObject(KOUNT_KEY));
@@ -169,10 +169,18 @@ public class Configuration {
     }
 
     /**
-     * @return instance of {@link AndroidPayConfiguration}.
+     * @return instance of {@link GooglePaymentConfiguration}.
      */
+    public GooglePaymentConfiguration getGooglePayment() {
+        return mGooglePaymentConfiguration;
+    }
+
+    /**
+     * @deprecated Use {@link #getGooglePayment()}.
+     */
+    @Deprecated
     public AndroidPayConfiguration getAndroidPay() {
-        return mAndroidPayConfiguration;
+        return getGooglePayment();
     }
 
     /**