+++ Demo/src/main/java/com/braintreepayments/demo/Settings.java
@@ -156,6 +156,10 @@ public class Settings {
         return getPreferences(context).getString("google_payment_currency", "USD");
     }
 
+    public static String getGooglePaymentMerchantId(Context context) {
+        return getPreferences(context).getString("google_payment_merchant_id", "18278000977346790994");
+    }
+
     public static List<String> getGooglePaymentAllowedCountriesForShipping(Context context) {
         String[] preference = getPreferences(context).getString("google_payment_allowed_countries_for_shipping", "US")
                 .split(",");
@@ -167,6 +171,9 @@ public class Settings {
         return countries;
     }
 
+    /**
+     * @deprecated Future Payments are deprecated
+     */
     public static boolean isPayPalAddressScopeRequested(Context context) {
         return getPreferences(context).getBoolean("paypal_request_address_scope", false);
     }